+++ examples/complete/main.tf
@@ -5,7 +5,7 @@ terraform {
   required_providers {
     artifactkeeper = {
       source  = "nicola-preda/artifact-keeper"
-      version = "~> 1.6.3"
+      version = "~> 1.6.4"
     }
   }
 }